pyproject.toml
[build-system]
requires = ["setuptools>=59.6.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = 'ais'
version = '1.0.0'
requires-python = '>=3.10'
dependencies = [
    "torchaudio>=2.4.1",
    "torchvision>=0.19.1",
    "scikit-learn>=1.5.2",
    "matplotlib>=3.9.2",
    "more-itertools>=10.5.0",
    "umap-learn>=0.5.6"]
license = { file = "LICENSE" }
authors = [{ name = 'Jose Llarena', email = 'jose.llarena@gmail.com' }]
maintainers = [{ name = 'Jose Llarena', email = 'jose.llarena@gmail.com' }]
description = "code for 'Automata in Space: Formal Language Theory meets Neural Computation' paper"
readme = "README.md"
keywords = ["neural networks", "spectral learning", "probing", "wfsa", "parity", "machine learning"]
classifiers = ['Development Status :: 5 - Production/Stable',
    'Environment :: Other Environment',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: MIT License',
    'Natural Language :: English',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Topic :: Scientific/Engineering :: Artificial Intelligence']

[project.urls]
repository = "https://github.com/JoseLlarena/ais"

[project.optional-dependencies]
dev = ["pytest>=8.3.3", "pytest-cov>=5.0.0"]
